requirements.txt
requests~=2.27.1
cachetools~=4.2.2
python-telegram-bot~=13.10